Refactor the application for data collection

+++ android/DataCollection/app/src/main/java/com/example/datacollection/TaskList.java
@@ -129,7 +129,7 @@ public class TaskList {
                 return id;
             }
 
-            public int getDuration() {
+            public long getDuration() {
                 return duration;
             }
